use common logic to a single macro module
